refactor: move service to internal
